User can edit question and delete answer

+++ app/controllers/answers_controller.rb
@@ -12,12 +12,7 @@ class AnswersController < ApplicationController
   end
 
   def destroy
-    if current_user.author?(answer)
-      answer.destroy
-      redirect_to answer.question, notice: 'Your answer successfully deleted.'
-    else
-      redirect_to answer.question
-    end
+    answer.destroy if current_user.author?(answer)
   end
 
   private
